backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
playwright==1.38.0
supabase==1.0.3
httpx==0.23.3
pydantic==1.10.8
python-dotenv==1.0.0